requirements.txt
pyperclip>=1.8.2
keyboard>=0.13.5; sys_platform == "win32"
pynput>=1.7.6; sys_platform != "win32"
pystray>=0.19.4
Pillow>=10.0.0
